Add code for RAFT
Add implementation for the algorithm described by the paper "Run Away
From your Teacher: Understanding BYOL by a Novel Self-Supervised
Approach" ( https://arxiv.org/abs/2011.10944 )

The RAFT class is essentially a copy-paste of the BYOL class, with
slight changes to the `forward` method, which computes a different loss
functionm making use of the new `raft_loss` function. RAFT's loss is the
difference of two losses: an "alignment loss" between the projection of
two different augmented views of the same image, to be minimized, and
the "cross-model loss", to be maximized, which is the distance between
the online and target representation of the same input, averaged over
the two different views.